# openapi.yaml
openapi: 3.0.3
info:
  title: LOOP API
  version: 1.0.0
  description: |
    LOOP Website API contract (public + admin).
    Response convention:
      - Success single: { data: T }
      - Success list: { data: T[], pagination }
      - Error: { error: string, code?: string }
servers:
  - url: https://loop.vn
    description: Production
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /api/search:
    get:
      tags: [Public]
      summary: Global search
      parameters:
        - in: query
          name: q
          schema: { type: string }
          required: true
        - in: query
          name: locale
          schema:
            type: string
            enum: [vi, en]
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessSearch'
...
components:
  schemas:
...
    SearchResult:
      type: object
      properties:
        services:
          type: array
          items: { type: object }
        team:
          type: array
          items: { type: object }
        projects:
          type: array
          items: { type: object }
        total: { type: integer }
...
    SuccessSearch:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SearchResult'
...